// archive/packages/dataforge-old/src/generated/client-entities.ts
export interface Task {
  id: any;
  createdAt: any;
  updatedAt: any;
  clientId?: any;
  title: any;
  description?: any;
  legacyStatus?: any;
  priority: any;
  dueDate?: any;
  startDate?: any;
  completedAt?: any;
  timeRange?: any;
  estimatedDuration?: any;
  legacyTags?: any;
  project?: any;
  assignee?: any;
  comments: any;
  tags: any;
}
// ...
export class Task implements Task {}
// ...
export const TaskStatus = {
  OPEN: 'open',
  IN_PROGRESS: 'in_progress',
  COMPLETED: 'completed',
} as const;
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];

export const TaskPriority = {
  LOW: 'low',
  MEDIUM: 'medium',
  HIGH: 'high',
} as const;
export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];